Potential fix for pull request finding 'CodeQL / Comparison of narrow type with wide type in loop condition'

+++ src/io/iobountytasks.cpp
@@ -203,7 +203,7 @@ void IOBountyTasks::fillMissingCreatures(const std::shared_ptr<Player> &player) 
 
 	auto maxIndex = static_cast<int32_t>(bestiaryList.size() - 1);
 	uint8_t creaturesGenerated = static_cast<uint8_t>(bountyData.currentCreaturesList.size());
-	uint8_t tries = 0;
+	uint16_t tries = 0;
 	while (creaturesGenerated < BOUNTY_MAX_CREATURES && tries < 1000) {
 		uint16_t raceId = (*(std::next(bestiaryList.begin(), uniform_random(0, maxIndex)))).first;
 		tries++;
